Added components named scripts
