+++ src/components/layout/Sidebar/index.js
@@ -1,21 +1,3 @@
-import React from "react"
+import Sidebar from "./Sidebar"
 
-import "./sidebar.scss"
-import MyChannel from "./MyChannel"
-import NewChannels from "./NewChannels"
-
-const Sidebar = () => {
-    return (
-        <aside className="sidebar">
-            <div className="sidenav">
-                <MyChannel />
-                <NewChannels />
-            </div>
-            <small className="footer-notice">
-                Copyright © {new Date().getFullYear()} etherna.io.
-            </small>
-        </aside>
-    )
-}
-
-export default React.memo(Sidebar)
+export default Sidebar
